Add files via upload
The build file is quite the size, so I haven't uploaded it. 
I was guided through the Arduino bit by @Kennedy Kolute.
I used Xampp as my local server, for the project.
I enjoyed working on this project and hope a beginner out there can use this as a resource, and even make it vast and better.